Extract editor playback render loop
